Update mail fetch functionality and dependencies
Expanded the mail fetch capacity from 10 to 100 emails and enhanced the email object to include id and body. Renamed 'mail' to 'mails' in the API route for conformity. Also updated the 'yarn.lock' file with new packages or versions. These changes enhance the mail search functionality and keep the dependencies up-to-date.